Refactoring to be able to add new commands via progrium/go-extpoints.
