+++ extpoints/interfaces.go
@@ -1,0 +1,7 @@
+package extpoints
+
+type CommandFunc func([]string)
+
+type CommandProvider interface {
+	Commands() map[string]CommandFunc
+}
